# Lazy load `networkx` in topology modules