--- a/openff/toolkit/topology/molecule.py
+++ b/openff/toolkit/topology/molecule.py
@@ -48,7 +48,6 @@ from typing import (
     overload,
 )
 
-import networkx as nx
 import numpy as np
 from openff.units.elements import MASSES, SYMBOLS
 from openff.utilities.exceptions import MissingOptionalDependencyError
@@ -92,6 +91,7 @@ from openff.toolkit.utils.utils import get_data_file_path, requires_package
 
 if TYPE_CHECKING:
     import IPython.display
+    import networkx as nx
     import nglview
     from rdkit.Chem import Mol as RDMol
 
@@ -1939,8 +1939,8 @@ class FrozenMolecule(Serializable):
 
     @staticmethod
     def are_isomorphic(
-        mol1: Union["FrozenMolecule", "_SimpleMolecule", nx.Graph],
-        mol2: Union["FrozenMolecule", "_SimpleMolecule", nx.Graph],
+        mol1: Union["FrozenMolecule", "_SimpleMolecule", "nx.Graph"],
+        mol2: Union["FrozenMolecule", "_SimpleMolecule", "nx.Graph"],
         return_atom_map: bool = False,
         aromatic_matching: bool = True,
         formal_charge_matching: bool = True,
@@ -2167,7 +2167,7 @@ class FrozenMolecule(Serializable):
 
     def is_isomorphic_with(
         self,
-        other: Union["FrozenMolecule", "_SimpleMolecule", nx.Graph],
+        other: Union["FrozenMolecule", "_SimpleMolecule", "nx.Graph"],
         **kwargs,
     ) -> bool:
         """
@@ -2785,7 +2785,7 @@ class FrozenMolecule(Serializable):
             if "_molecule_atom_index" in atom.__dict__:
                 del atom.__dict__["_molecule_atom_index"]
 
-    def to_networkx(self) -> nx.Graph:
+    def to_networkx(self) -> "nx.Graph":
         """Generate a NetworkX undirected graph from the molecule.
 
         Nodes are Atoms labeled with atom indices and atomic elements (via the ``element`` node atrribute).
@@ -3541,7 +3541,7 @@ class FrozenMolecule(Serializable):
         return self._hill_formula
 
     @staticmethod
-    def _object_to_hill_formula(obj: Union["FrozenMolecule", nx.Graph]) -> str:
+    def _object_to_hill_formula(obj: Union["FrozenMolecule", "nx.Graph"]) -> str:
         """Take a Molecule or NetworkX graph and generate its Hill formula.
         This provides a backdoor to the old functionality of Molecule.to_hill_formula, which
         was a static method that duck-typed inputs of Molecule or graph objects."""
@@ -5730,7 +5730,7 @@ class Molecule(FrozenMolecule):
             pass
 
 
-def _networkx_graph_to_hill_formula(graph: nx.Graph) -> str:
+def _networkx_graph_to_hill_formula(graph: "nx.Graph") -> str:
     """
     Convert a NetworkX graph to a Hill formula.
 
@@ -5810,6 +5810,8 @@ def _nth_degree_neighbors_from_graphlike(
     neighbors
         tuples (len 2) of atom that are separated by ``n`` bonds.
     """
+    import networkx as nx
+
     graph = graphlike.to_networkx()
 
     for node_i in graph.nodes: